packages/aevum-sdk/pyproject.toml
[project]
name = "aevum-sdk"
version = "0.1.0"
description = "Aevum — complication developer kit. Build extensions in under 30 minutes."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "aevum-core",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-api>=1.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.10",
    "ruff>=0.9",
]

[project.scripts]
aevum = "aevum.sdk.__main__:main"

[project.urls]
Homepage = "https://aevum.build"
Repository = "https://github.com/aevum-labs/aevum"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aevum"]

[tool.uv.sources]
aevum-core = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--tb=short"
pythonpath = ["src", "tests"]

[tool.mypy]
strict = true
python_version = "3.11"
mypy_path = "src"
explicit_package_bases = true
ignore_missing_imports = true

[tool.ruff]
line-length = 130

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I", "ANN"]
ignore = ["ANN401"]
